% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{print.SPRE}
\alias{print.SPRE}
\title{print SPRE}
\usage{
\method{print}{SPRE}(x, ...)
}
\arguments{
\item{x}{an object of class 'SPRE'}

\item{...}{arguments passed to print.default}
}
\value{
output some information to console
}
\description{
print SPRE
}
\details{
print an SPRE object
}
\examples{
data('HEAT_stat')
modelfit<-SPRE(predictor=HEAT_stat$Session,response=HEAT_stat$FData)
print(modelfit)
}
\references{
Weissman-Miller, D., Shotwell, M.P. & Miller, R.J. (2012).  New single-subject and small-n design in occupational therapy:  Application to weight loss in obesity.  American Journal of Occupational Therapy, 66, 455-462.
Weissman-Miller, D. (2013). Novel point estimation from a Semiparametric Ratio Estimator (SPRE): Long-term health outcomes from short-term linear data, with application to weight loss in obesity. International Journal of Biostatistics, 9(2): 175-184.
Weissman-Miller, D., & Graham, K. C. (2015). Novel scale development for fear of falling and falls: Analysed using a Semiparametric Ratio Estimator (SPRE). International Journal of Statistics and Probability, 4(3), 161.
Weissman-Miller, D. (2016). On predicting survival in prostate cancer: using an extended maximum spacing method at the change point of the semiparametric ratio estimator (SPRE). International Journal of Statistics and Probability, 5(2), 19.
Weissman-Miller, D., Miller, R. J., & Shotwell, M. P. (2017). Translational Research for Occupational Therapy: Using SPRE in Hippotherapy for Children with Developmental Disabilities. Occupational therapy international, 2017.
}
\author{
Deborah Weissman-Miller
}
